Added Aeon::Room and Aeon::Character classes.
Players have a 1-to-1 association with a Character. I'm thinking of the Player object acting as an "account" for the human player, as well as where all the interface logic is stored like interpreting commands. The player's Character is just the player's embodiment in the game world, and will likely inherit from the same class as NPC's or actually any "game object".

Since this is all new to me, so far my motto is "go for it until it sucks". :)